Add adaptive plot step and animation interval scaling
Could be made better be reevaluating whether the browser starts to do
better as time goes on (speculative optimization kicks in? or is that
just optimistic...)

At any rate, some data for comparison:

On my system, FireFox 4 settles out at a 32 step and 150-300ms interval,
where-as Chrome is able to not scale back what-so-ever... I am
unimpressed with FailFox.

I originally tried just scaling the step thinking "surely firefox only
needs a step or two"... however it scaled all the way back to 256 step
and looked like something from an 8-bit system...